Change to hibernate 4 jars

+++ src/main/java/org/candlepin/model/AbstractHibernateCurator.java
@@ -28,7 +28,7 @@ import org.hibernate.criterion.Order;
 import org.hibernate.criterion.Projection;
 import org.hibernate.criterion.Projections;
 import org.hibernate.criterion.Restrictions;
-import org.hibernate.impl.CriteriaImpl;
+import org.hibernate.internal.CriteriaImpl;
 import org.hibernate.transform.ResultTransformer;
 
 import java.io.Serializable;
@@ -173,7 +173,7 @@ public abstract class AbstractHibernateCurator<E extends Persisted> {
 
     private Integer findRowCount(Criteria c) {
         c.setProjection(Projections.rowCount());
-        return (Integer) c.uniqueResult();
+        return ((Long) c.uniqueResult()).intValue();
     }
 
     @SuppressWarnings("unchecked")
